+++ index.html
@@ -11,7 +11,7 @@
 </head>
 
 <body>
-  <script src="/js/main.js">
+  <script src="/js/main.js" type="module">
   </script>
   <main>
