+++ js/constants.js
@@ -1,0 +1,37 @@
+export const NAMES = [
+  'Роман',
+  'Алексей',
+  'Татьяна',
+  'Илья',
+  'Галина',
+  'Дмитрий',
+  'Анна',
+  'Андрей',
+  'Мария',
+  'Николай'
+];
+export const DESCRIPTION = [
+  'Крутая фота',
+  'Просто улет',
+  'Отвал башки',
+  'Умопомрачительно',
+  'Фи, какая гадость',
+  'Могло быть и хуже',
+  'Все в сад',
+  'Бывают в жизни огорчения',
+  'Превосходная фота',
+  'Авации и аплодисменты'
+];
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const MAX_PHOTO_COMMENTS = 30;
+export const MAX_NUMBER_PUBLISHED_PHOTOS = 25;
+export const MIN_NUMBER_LIKES = 15;
+export const MAX_NUMBER_LIKES = 200;
